+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Bug-069970 - On product detail page option vaules should be displated as per display order
 - Bug-070009 - On deleting seller product, its options are not deleting.
 - Bug-070036 - Admin top selling products should be based on completed orders.
+- Bug-070110 - User dashboard area breadcrumb labels shall be manageable from admin end
 # RV-10.0.0.20221123
 
 ## Bug Fixes: